BETA: Fix for: Couple actions works fine from player A to B, but when couple finished and player B tries to request couple to player A it says "player is already doing a couple action always", until players A re-login. 	* Reported and tested by MELERIX.
The problem:
Player A request player B.
Player B answer, player A's "couple data" is cleaned only if player B cancel.
So when couple action is accepted player A has player B as target of couple action.
The fix:
"Clean" (set to 0) the player A's "couple data" for any answer from player B.